requirements.txt
streamlit==1.32.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
openai==1.12.0
pandas==2.2.1
ctransformers==0.2.27
ollama==0.1.6 